pyproject.toml
[tool.poetry]
name = "djangoprojectdoggystyle"
version = "0.1.0"
description = ""
authors = ["Svennson17 <kiryl.shvedau@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.6"
django-bootstrap4 = "^22.1"
static = "^1.1.1"
dj-config-url = "^0.1.1"
gunicorn = "^20.1.0"
dj-database-url = "^1.0.0"
psycopg2 = "^2.9.3"
Pillow = "^9.2.0"
whitenoise = "^6.2.0"
djangorestframework = "^3.13.1"
generics = "^3.9.0"
response = "^0.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
